pyproject.toml
[project]
name = "sc3k"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "open3d>=0.19.0",
    "rerun-sdk>=0.22.1",
    "seaborn>=0.13.2",
    "tensorboard>=2.14.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
]
